+++ apps/dash-uber-rides-demo/constants.py
@@ -1,0 +1,52 @@
+import pandas as pd
+import numpy as np
+
+# Plotly mapbox public token
+mapbox_access_token = "pk.eyJ1IjoicGxvdGx5bWFwYm94IiwiYSI6ImNrOWJqb2F4djBnMjEzbG50amg0dnJieG4ifQ.Zme1-Uzoi75IaFbieBDl3A"
+
+# Dictionary of important locations in New York
+list_of_locations = {
+    "Madison Square Garden": {"lat": 40.7505, "lon": -73.9934},
+    "Yankee Stadium": {"lat": 40.8296, "lon": -73.9262},
+    "Empire State Building": {"lat": 40.7484, "lon": -73.9857},
+    "New York Stock Exchange": {"lat": 40.7069, "lon": -74.0113},
+    "JFK Airport": {"lat": 40.644987, "lon": -73.785607},
+    "Grand Central Station": {"lat": 40.7527, "lon": -73.9772},
+    "Times Square": {"lat": 40.7589, "lon": -73.9851},
+    "Columbia University": {"lat": 40.8075, "lon": -73.9626},
+    "United Nations HQ": {"lat": 40.7489, "lon": -73.9680},
+}
+
+# Initialize data frame
+df1 = pd.read_csv(
+    "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data1.csv",
+    dtype=object,
+)
+df2 = pd.read_csv(
+    "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data2.csv",
+    dtype=object,
+)
+df3 = pd.read_csv(
+    "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data3.csv",
+    dtype=object,
+)
+df = pd.concat([df1, df2, df3], axis=0)
+df["Date/Time"] = pd.to_datetime(df["Date/Time"], format="%Y-%m-%d %H:%M")
+df.index = df["Date/Time"]
+df.drop("Date/Time", 1, inplace=True)
+totalList = []
+for month in df.groupby(df.index.month):
+    dailyList = []
+    for day in month[1].groupby(month[1].index.day):
+        dailyList.append(day[1])
+    totalList.append(dailyList)
+totalList = np.array(totalList)
+
+
+
+# Gets the amount of days in the specified month
+# Index represents month (0 is April, 1 is May, ... etc.)
+daysInMonth = [30, 31, 30, 31, 31, 30]
+
+# Get index for the specified month in the dataframe
+monthIndex = pd.Index(["Apr", "May", "June", "July", "Aug", "Sept"])
