+++ apps/dash-uber-rides-demo/requirements.txt
@@ -1,3 +1,3 @@
 dash==1.12.0
-pandas==0.24.1
-gunicorn==19.9.0
+pandas==1.4.2
+gunicorn==20.1.0
