+++ apps/dash-uber-rides-demo/utils/figures.py
@@ -1,0 +1,185 @@
+from datetime import datetime as dt
+from plotly import graph_objs as go
+import numpy as np
+
+from utils.helper_functions import get_selection, getLatLonColor
+from constants import list_of_locations, mapbox_access_token
+
+def map(date_picked, bars_selected, location):
+    zoom = 12.0
+    latInitial = 40.7272
+    lonInitial = -73.991251
+    bearing = 0
+
+    if location:
+        zoom = 15.0
+        latInitial = list_of_locations[location]["lat"]
+        lonInitial = list_of_locations[location]["lon"]
+
+    date_picked = dt.strptime(date_picked, "%Y-%m-%d")
+    monthPicked = date_picked.month - 4
+    dayPicked = date_picked.day - 1
+    listCoords = getLatLonColor(bars_selected, monthPicked, dayPicked)
+
+    return go.Figure(
+        data=[
+            # Data for all rides based on date and time
+            go.Scattermapbox(
+                lat=listCoords["Lat"],
+                lon=listCoords["Lon"],
+                mode="markers",
+                hoverinfo="lat+lon+text",
+                text=listCoords.index.hour,
+                marker=dict(
+                    showscale=True,
+                    color=np.append(np.insert(listCoords.index.hour, 0, 0), 23),
+                    opacity=0.5,
+                    size=5,
+                    colorscale=[
+                        [0, "#F4EC15"],
+                        [0.04167, "#DAF017"],
+                        [0.0833, "#BBEC19"],
+                        [0.125, "#9DE81B"],
+                        [0.1667, "#80E41D"],
+                        [0.2083, "#66E01F"],
+                        [0.25, "#4CDC20"],
+                        [0.292, "#34D822"],
+                        [0.333, "#24D249"],
+                        [0.375, "#25D042"],
+                        [0.4167, "#26CC58"],
+                        [0.4583, "#28C86D"],
+                        [0.50, "#29C481"],
+                        [0.54167, "#2AC093"],
+                        [0.5833, "#2BBCA4"],
+                        [1.0, "#613099"],
+                    ],
+                    colorbar=dict(
+                        title="Time of<br>Day",
+                        x=0.93,
+                        xpad=0,
+                        nticks=24,
+                        tickfont=dict(color="#d8d8d8"),
+                        titlefont=dict(color="#d8d8d8"),
+                        thicknessmode="pixels",
+                    ),
+                ),
+            ),
+            # Plot of important locations on the map
+            go.Scattermapbox(
+                lat=[list_of_locations[i]["lat"] for i in list_of_locations],
+                lon=[list_of_locations[i]["lon"] for i in list_of_locations],
+                mode="markers",
+                hoverinfo="text",
+                text=[i for i in list_of_locations],
+                marker=dict(size=8, color="#ffa0a0"),
+            ),
+        ],
+        layout=go.Layout(
+            autosize=True,
+            margin=go.layout.Margin(l=0, r=35, t=0, b=0),
+            showlegend=False,
+            mapbox=dict(
+                accesstoken=mapbox_access_token,
+                center=dict(lat=latInitial, lon=lonInitial),  # 40.7272  # -73.991251
+                style="dark",
+                bearing=bearing,
+                zoom=zoom,
+            ),
+            updatemenus=[
+                dict(
+                    buttons=(
+                        [
+                            dict(
+                                args=[
+                                    {
+                                        "mapbox.zoom": 12,
+                                        "mapbox.center.lon": "-73.991251",
+                                        "mapbox.center.lat": "40.7272",
+                                        "mapbox.bearing": 0,
+                                        "mapbox.style": "dark",
+                                    }
+                                ],
+                                label="Reset Zoom",
+                                method="relayout",
+                            )
+                        ]
+                    ),
+                    direction="left",
+                    pad={"r": 0, "t": 0, "b": 0, "l": 0},
+                    showactive=False,
+                    type="buttons",
+                    x=0.45,
+                    y=0.02,
+                    xanchor="left",
+                    yanchor="bottom",
+                    bgcolor="#323130",
+                    borderwidth=1,
+                    bordercolor="#6d6d6d",
+                    font=dict(color="#FFFFFF"),
+                )
+            ],
+        ),
+    )
+
+
+def histogram(date_picked, bars_selected):
+    date_picked = dt.strptime(date_picked, "%Y-%m-%d")
+    monthPicked = date_picked.month - 4
+    dayPicked = date_picked.day - 1
+
+    [xVal, yVal, colorVal] = get_selection(monthPicked, dayPicked, bars_selected)
+
+    layout = go.Layout(
+        bargap=0.01,
+        bargroupgap=0,
+        barmode="group",
+        margin=go.layout.Margin(l=10, r=0, t=0, b=50),
+        showlegend=False,
+        plot_bgcolor="#323130",
+        paper_bgcolor="#323130",
+        dragmode="select",
+        font=dict(color="white"),
+        xaxis=dict(
+            range=[-0.5, 23.5],
+            showgrid=False,
+            nticks=25,
+            fixedrange=True,
+            ticksuffix=":00",
+        ),
+        yaxis=dict(
+            range=[0, max(yVal) + max(yVal) / 4],
+            showticklabels=False,
+            showgrid=False,
+            fixedrange=True,
+            rangemode="nonnegative",
+            zeroline=False,
+        ),
+        annotations=[
+            dict(
+                x=xi,
+                y=yi,
+                text=str(yi),
+                xanchor="center",
+                yanchor="bottom",
+                showarrow=False,
+                font=dict(color="white"),
+            )
+            for xi, yi in zip(xVal, yVal)
+        ],
+    )
+
+    return go.Figure(
+        data=[
+            go.Bar(x=xVal, y=yVal, marker=dict(color=colorVal), hoverinfo="x"),
+            go.Scatter(
+                opacity=0,
+                x=xVal,
+                y=yVal / 2,
+                hoverinfo="none",
+                mode="markers",
+                marker=dict(color="rgb(66, 134, 244, 0)", symbol="square", size=40),
+                visible=True,
+            ),
+        ],
+        layout=layout,
+    )
